refactor: add image fieldType
