+++ packages/visual-editor/src/editor/YextField.test.tsx
@@ -5,7 +5,6 @@ import { msg } from "../utils/i18n/platform.ts";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import { YextField } from "./YextField.tsx";
 import { VIDEO_CONSTANT_CONFIG } from "../internal/puck/constant-value-fields/Video.tsx";
-import { IMAGE_CONSTANT_CONFIG } from "../internal/puck/constant-value-fields/Image.tsx";
 
 const {
   dynamicOptionsSelectorMock,
@@ -226,20 +225,6 @@ describe("YextField", () => {
       true
     );
     expect(field).toBe(returnedField);
-  });
-
-  it("returns image configs with the provided label", () => {
-    const fieldName = msg("fields.image", "Image");
-
-    const field = YextField(fieldName, {
-      type: "image",
-    }) as any;
-
-    expect(field).toMatchObject({
-      type: IMAGE_CONSTANT_CONFIG.type,
-      label: fieldName,
-    });
-    expect(field.render).toBe(IMAGE_CONSTANT_CONFIG.render);
   });
 
   it("returns video configs with the provided label", () => {
