+++ packages/visual-editor/src/fields/fields.ts
@@ -24,6 +24,7 @@ import {
   OptionalNumberField,
   OptionalNumberFieldOverride,
 } from "./OptionalNumberField.tsx";
+import { ImageField, ImageFieldOverride } from "./ImageField.tsx";
 
 export type YextPuckFields = {
   basicSelector: BasicSelectorField;
@@ -31,6 +32,7 @@ export type YextPuckFields = {
   code: CodeField;
   dateTimeSelector: DateTimeSelectorField;
   fontSizeSelector: FontSizeSelectorField;
+  image: ImageField;
   optionalNumber: OptionalNumberField;
 };
 
@@ -59,5 +61,6 @@ export const YextPuckFieldOverrides = {
   code: CodeFieldOverride,
   dateTimeSelector: DateTimeSelectorFieldOverride,
   fontSizeSelector: FontSizeSelectorFieldOverride,
+  image: ImageFieldOverride,
   optionalNumber: OptionalNumberFieldOverride,
 };
