+++ packages/visual-editor/src/fields/index.ts
@@ -10,6 +10,7 @@ export {
 } from "./CTASelectorField.tsx";
 
 export { CodeFieldOverride, type CodeField } from "./CodeField.tsx";
+export { ImageFieldOverride, type ImageField } from "./ImageField.tsx";
 
 export {
   DateTimeSelectorFieldOverride,
